backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart

#running backend:- uvicorn main:app --reload
#venv setup: .\venv\Scripts\Activate.ps1